CX: Nominate a featured collection via CommunityConfiguration
This patch adds the CommunityConfiguration extension as a dependency
and uses it to allow communities to set a local featured collection.
This featured collection is exposed in two ways:

* Via a new API endpoint `cxconfig`: This will be useful when the
target language is different then the current wiki langauge.

* And also provided as a JS config variable. This will be useful
when the target language is the same as the current wiki language.

For testing locally, use the CommunityConfiguration extension or set
$wgContentTranslationFeaturedCollection = 'Climate';
in your local settings.

Bug: T408625
Change-Id: I31479568855e36d70bde34b83e139e7205702249