CX: Fix easy article suggestions for newcomers
When the current user is a newcomer, we want to display "easy"
recommendations to them. Moreover, if the current device is mobile, easy
article suggestions shown to mobile users must be based on the lead
section size. However, currently the thresholds that are sent to the
article suggestions API when on mobile, are the article thresholds,
instead of the section thresholds.

This patch fixes this issue by sending the proper thresholds to the API
for easy recommendations on mobile. At the same time, it fixes the
"Quick translation" indicator on mobile, to be displayed only if the
lead section translation is an easy translation.

Bug: T403730
Change-Id: I648be6b40aebe2350b4d73b498ce9b80c1be9e66