CX edit-based suggestions: Shuffle seeds to ensure random seed is used
Currently, when edit-based suggestions are fetched from the
Recommendation API, it is always the same seed that is used: that is
the last published translation by the current user.

To ensure that edit-based suggestions are diverse and not always
limited to the same topic, this patch uses the "shuffleArray" helper
function to make sure that the seeds array is randomly ordered every
time.

Change-Id: I8714a7c45356567064b953054d60c30be804b8c0