Fetch supported languages via sitematrix
This is a partial revert of Icce4af122784832989e228eb792261d3baf75eca
that replaced CX server's language pairs API with a hardcoded list.

This patch restores useApplicationLanguagesInitialize and
useSupportedLanguageCodes similar to how they were but uses the
sitematrix API from meta to get the list of language codes
for which there is a Wikipedia site that is not closed.

Bug: T409481
Change-Id: Ifd5051c64ceef1f6cadc06f8dd9e9613d484ee7f