Deduplicate article search query
useArticleSearch is used in both SXArticleSearch.vue and
SearchResultsCard.vue. As a result, when the input changes, the
composable triggers the search from both contexts.

This patch keeps useArticleSearch only in the parent component
(SXArticleSearch.vue) and passes down the results as props to
the child component (SearchResultsCard.vue).

Also move the search tests from the UI component to the
composable.

Change-Id: I39f79ca04179ecdbc7e9e8feb9e5098407757c2c