Fix parameter order in collections suggestions functions
The functions to fetch article and section sugestions were
recently modified in If4edb82e12052aa742d43108e834e5950e22faad
to accomodate the continue_seed and continue_offset parameter.

However, the arguments of the two functions used to get
collection-based suggestions were reorder and the code ended
up putting the name of the collection in the count api
parameter.

This commit re-orders the function parameter.

As a bonus, extract the count=24 hardcoded everywhere into
a DEFAULT_COUNT constant and reduce it to 6.

Change-Id: I2704692a2acbd67195a60dc20eaef84776253c8b