CX: Over represent featured collection in "all collections" filter
This patch adds the "featured_collection" URL parameter to the request
URL, when "All collections" filter is selected and featured collection
is set. With this URL parameter filled, the 50% of the returned
recommendations comes from the featured collection.

This patch will not produce any changes in CX until the feature is
deployed in the Recommendation API.

Bug: T409332
Change-Id: Ia88732b37984ecf60752536c85ac2749fad47998