CX: Pass the actual collection name to next unseen suggestion getter
For some reason, until now instead of passing the name of the collection
to the `getNextUnseenSuggestionByCollection` function, an array of
section suggestions was passed. That made the return value of the
`getNextUnseenSectionSuggestionByCollection` function to always be null,
and thus a new request to fetch featured collection section suggestions
was sent every time the refresh button was clicked. This is now fixed.

Change-Id: I2da3ac451d4c32af358b62494de08bccd01ae029