CX:  Override parent's onTagSelect for category target language
Selected tags don't have corresponding menu items, so trying to scroll
them causes null reference errors.

Clicking on the target language category in the CX editor should not
have an effect. Therefore, we override the onTagSelect method as it
throws a TypeError and introduces noise in logstash as well. All other
initial functionality is preserved.

Bug: T410366
Change-Id: I7fa6a651db5b45709f9df1c919095fc48c9af5ca