Compare contents: Always hide secondary "Translate" button on route load
When the SXContentComparatorContentHeader SFC is mounted, we create and
start a new intersection observer to check whether the content header
is fully visible. If the content header is not fully visible, we should
make it sticky and also display the secondary "Translate" button inside
it. However, on tablet screens (and even in some cases on desktop),
the height of the intersection rectangle is found to be 0, during
component mount. This can happen because the navigation transition has
not yet been completed and the elements are not properly positioned at
the time of the execution of the observer callback.

To fix this issue, this patch starts the intersection observer on the
first time that the content is scrolled, to make sure that the
callback is executed only after the elements are properly rendered and
positioned.

Bug: T413258
Change-Id: Id39c94307f4fd7ee6519e3c7e1d0807adf84c6b7